Set `fail_on` to `methods` in SourceClear scan
`high` is problematic for us because it is active if you use a vulnerable method or not. In addition, the SourceClear database seems to file a vulnerability wrong, which makes one of our builds fail without good reason. Since we don't use that method, the build won't fail anymore once this is merged.

More information on how this setting works is available at [1].

1: https://www.sourceclear.com/blog/Vulnerable-Methods-Under-the-Hood/