Check deployment descriptors in quality checks
In case of a productive branch and unsafeMode is false the pipeline will check if there are prohibited environment variables set in any of the deployment descriptors. If prohibited variables are set the pipeline will log it as a warning and pin badges to the pipeline and the pipeline summary.