Allow user to define a ci-lint script which is executed as part of the lint stage
This change allows the user to provide a custom linting script to be
executed as part of the lint stage of the pipeline. In addition, the
default behavior of the lint stage is extended to not only provide the
possibility to run the UI5 lint step for UI5 projects, but also provide
an execution of ESLint with a provided default configuration for non-SAPUI5
Javascript/Typescript projects.

Previously, the lint stage just executed the UI5 lint step in case a SAPUI5
project could be identified or the stage was skipped otherwise. The only
way to run a custom linter was to make use of the extension mechanism.
Now it is possible to run a custom linter by defining a script named
'ci-lint' in the 'scripts' section in the 'package.json' file or files of the 
project. In addition, the output of the linting tool is expected to be in
checkstyle XML format and the file is name '*cilint.xml', which may reside
in any directory of the project.

If no ci-lint script is defined and the project is not a SAP UI5 project,
ESLint is executed with a default configuration, which is applicable to
a variety of Javascript/Typescript projects.