Fix lint stage skipping
This change fixes the lint stage skip behavior in case hidden JS or TS
files are present in the project. Previously, the lint stage was wrongly
executed if the project contained hidden JS files. By default, ESLint
does not lint hidden JS files and as a result if only hidden JS files
are present in the project, the lint stage is still executed but no files
are linted.
Now, the lint stage is skipped in case only hidden files are present in
the project.