createHdiContainer: Replace dir() call in initializeDatabase()
This change replaces the dir() call with passing the path of the db module
via buildDescriptorList parameter to npmExecuteScripts step. Previously,
when changing into the db module directory for the execution of the npm
script, the npmExecuteScripts step again unstashed the sources stash
as defined in its metadata. In addition, the npmExecuteScripts step was only
able to execute the configured script for all package.json files it can find
that contain this script. Due to the unstash of the sources stash inside the db
module directory the npmExecuteScripts step possibly found multiple
package.json files containing the script start leading to unintended behavior.

Now, the npmExecuteScripts step is able to execute a script only for the
modules configured via buildDescriptorList instead of for all modules
that implement the script.